Add Funtions To Post New Scores To API
