+++ src/index.html
@@ -3,9 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Leaderboard</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.3/font/bootstrap-icons.css">
-    <!-- <link rel="stylesheet" href="index.css" /> -->
+    <title>Chill Out Leaderboard!</title>
   </head>
   <body>
     <header>
