+++ src/modules/api_manager.js
@@ -1,5 +1,20 @@
-const reqScoreURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/Ev5yMffto11bWoMTkoDt/scores/';
+const reqScoreURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/aDlDXb9nUEqfbts6gGH9/scores/';
 const ScoreURL = new Request(reqScoreURL);
+
+const regdGame = async () => {
+  const requestURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/';
+  const request = new Request(requestURL);
+  const myGame = { name: 'Chill Out' };
+  fetch(request, {
+    method: 'POST',
+    body: JSON.stringify(myGame),
+    headers: {
+      'Content-type': 'application/json; charset=UTF-8',
+    },
+  })
+    .then((response) => response.json())
+    .then((json) => console.log(json));
+};
 
 const loadLeaderBoard = async (callBack) => {
   const response = await fetch(ScoreURL);
@@ -7,4 +22,18 @@ const loadLeaderBoard = async (callBack) => {
   callBack(leaderBoard.result);
 };
 
-export default loadLeaderBoard;
+const postNewScore = async (scoreData) => {
+  fetch(ScoreURL, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(scoreData),
+  })
+    .then((response) => response.json())
+    .then((score) => {
+      console.log('New score added:', score);
+    });
+};
+
+export { loadLeaderBoard, postNewScore, regdGame };
